process,syscall: delete POSIX dead trees
Drop process/{accounting,acct_record,clone_flags,nox,numa,operations,
realtime}/ and process/signal/syscall/. Drop syscall/{signals,tls}/.
None of these were on the active SyscallNumber dispatch chain; the
POSIX scaffold was reachable only from its own tests.

process::signal retains constants + send + delivery for kernel-internal
use (alarm tick keeps SIGALRM via process::signal::send_signal).
syscall surface is now Mk*/Crypto*/Admin*/Debug*/Graphics* only.

+++ src/process/alarm.rs
@@ -18,16 +18,14 @@
 //! The walk lives here in `process::` so the timer IRQ owns its own
 //! delivery path and does not reach across into a syscall module.
 
-use crate::syscall::signals::constants::SIGALRM;
-use crate::syscall::signals::delivery::send_signal;
+use crate::process::signal::{send_signal, SIGALRM};
 
-/// Walk the process table and deliver `SIGALRM` to any PCB whose
-/// alarm timestamp has expired. Intended to be called from the
-/// kernel timer IRQ tick.
+// Walk the process table and deliver SIGALRM to any PCB whose alarm
+// timestamp has expired. Called from the kernel timer IRQ tick.
 pub fn tick() {
     for pcb in crate::process::get_process_table().get_all_processes() {
         if pcb.check_alarm_expired() {
-            let _ = send_signal(pcb.pid, SIGALRM);
+            let _ = send_signal(pcb.pid, SIGALRM as u32);
         }
     }
 }
